Update to include a working version of the pixel growth command for calculating distances from regions defined with pixel values of 0.

+++ src/img/RSGISCalcDist2Geom.h
@@ -78,7 +78,7 @@ namespace rsgis{namespace img{
         void incrementCounter();
 		~RSGISCalcDistViaIterativeGrowth();
     protected:
-        unsigned int counter;
+        float counter;
         bool change;
         double imgRes;
 	};
